zip/africachange_backup_20251212T201945Z/requirements.txt
alembic==1.14.1
blinker==1.8.2
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.1.8
colorama==0.4.6
dnspython==2.6.1

email_validator==2.2.0
et_xmlfile==2.0.0
Flask==2.3.2
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.2.2
greenlet==3.1.1
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.5
Mako==1.3.9
MarkupSafe==2.1.5
mailjet-rest==1.3.4
openpyxl==3.1.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
python-dotenv
requests==2.32.4
six==1.17.0
SQLAlchemy==2.0.38
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.2.3
Werkzeug==2.3.7
zipp==3.20.2
psycopg2-binary
psycopg[binary]

gunicorn==21.2.0
setuptools
wheel
